# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Go Messenger API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Авторизация пользователя",
                "parameters": [
                    {
                        "description": "Учётные данные",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/project_internal_presentation_schemas.AuthorizeSchema"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/project_internal_presentation_schemas.TokensSchema"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/project_internal_presentation_schemas.ErrorSchema"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/project_internal_presentation_schemas.ErrorSchema"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Выдаёт новый access-токен по refresh-токену из cookie",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Обновление access-токена",
                "parameters": [
                    {
                        "type": "string",
                        "default": "refresh_token=\u003ctoken\u003e",
                        "description": "Refresh-токен в cookie",
                        "name": "refresh_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/project_internal_presentation_schemas.TokensSchema"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/project_internal_presentation_schemas.ErrorSchema"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/project_internal_presentation_schemas.ErrorSchema"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация пользователя",
                "parameters": [
                    {
                        "description": "Данные пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/project_internal_presentation_schemas.CreateUserSchema"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/project_internal_presentation_schemas.UserCreatedSchema"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/project_internal_presentation_schemas.ErrorSchema"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "project_internal_presentation_schemas.AuthorizeSchema": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "1Q2w3e"
                },
                "username": {
                    "type": "string",
                    "example": "maximEZ"
                }
            }
        },
        "project_internal_presentation_schemas.CreateUserSchema": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "maxim@mail.ru"
                },
                "password": {
                    "type": "string",
                    "example": "1Q2w3e"
                },
                "username": {
                    "type": "string",
                    "example": "maximEZ"
                }
            }
        },
        "project_internal_presentation_schemas.ErrorSchema": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "project_internal_presentation_schemas.TokensSchema": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                }
            }
        },
        "project_internal_presentation_schemas.UserCreatedSchema": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        }
    }
}